Use `AnnotateSerialize` as the object-safe trait for serialization
The idea is that when we move away from specialization, not all `Serialize`
type would implement `Annotate` anymore; therefore we split this part out
which can be a dedicated trait that is still implemented on all `Serialize` types.

One unit test is broken at this stage -- it'll be fixed in the upcoming commit.